Use commit:$REAL_COMMIT while creating reviews
A gerrit search is made by rdo_release_review for what is
specified against -n option of rdo_release_review.
Gerrit based on the value(-n) tries to find the meaning([1])
and does the generic search and it might fail to return the
valid result.
Since we are searching for a commit, we should do
a strict check by specifying -n commit:$REAL_COMMIT.

[1] https://review.openstack.org/Documentation/user-search.html

Depends-On: If3aa14ca946ecd199b49073cac166d89af383533
Change-Id: Ic7f07ffd35c518215d4c1d86334597298d914b90