Move the DLRN-rpmbuild-fedora job to the check pipeline
The job has already proved itself to be stable and useful (see [1][2]
for some examples). Let's move it to the check pipeline for all
projects. It remains as non-voting, since we don't want to hold
changes if there is something wrong at the Fedora Rawhide level.

Since having it in the check pipeline will increase its load, we are
increasing the min-servers in Nodepool to 2.

[1] - https://review.rdoproject.org/r/10537
[2] - https://review.rdoproject.org/r/10541

Change-Id: Iaa9c5f591ed140ff610df548b0f91ceb7bd0449b