Reduce max-servers for TripleO to 75
We suspect that Jenkins is having a hard time keeping up with the
concurrency we're imposing it right now (>90 concurrent slaves) which
is made problematic with slaves disconnecting, causing jobs to fail.

Let's try to reduce the concurrency a bit see if it helps while keeping
the Zuul queue reasonable.

Change-Id: I2a7f68f34fc1a2b33044a56f32e142425ae0a254