[WIP] centos-opstools checks/creates package modified
Currently the creation and the validation of the rpm
packages was based on the name of the project or on a
directory spec/*.spec where any spec file can be.

Everything works perfectly fine when the spec file is
named baseontheproject.spec but it doesn't when the spec
file is under the spec directory due to gating_scripts
jobs.

This aims to fix the rpmbuild and rpmlint jobs so the
spec file can be name baseproject.spec or
any spec file on the root directory.

Change-Id: Ib8f144bff7cfe6cd5ab602bd1300eece95394bc1