Update patch set 4
Patch Set 4:

I might be wrong, but so far the repositories only have one spec file Currently the specfile needs to have the name of the package that is going to build. So at this point we can only build packages that has the name of the repo. This patch only add the posibility of build rpm packages where the spec file does not have anything to do with the name of the repo.

We can make the script fails if there are 2 or more spec files, or we can make the script process all the spec files of the repositories.

Patch-set: 4
Label: Code-Review=0