Add OVN job (scenario007) to periodic pipeline
OVN is becoming more important so it may be good to include it
in periodic pipeline for master.

This patch adds job for featureset030 (based on scenario007) which
deploys neutron with OVN.

Note this patch is not adding this job as voting for promotion, which
will be a separated patch in ci-config repo once it's stabilized.

Change-Id: Ie7c2ca85f5d8ee3133e615d6a60698c027fd88ce