Update patch set 4
Patch Set 4: Code-Review-1

I think we're not following the right process here.

If we want to add a package to RDO, we should start by creating a package review bug in Bugzilla, then opening a review to rdoinfo, see steps 1 and 2 in https://www.rdoproject.org/documentation/add-packages/ . I have not found the rdoinfo review, please correct me if I'm wrong.

The review for the config repo will be created automatically after merging the rdoinfo review, so this one is not needed.

Patch-set: 4