Update patch set 7
Patch Set 7: Code-Review+2

Please keep in mind this is not adding any Zuul configuration (so no CI jobs on changes to the repo) or Gerrit replication (so changes in this repo will not be replicated to github.com/redhat-openstack/whitebox-tempest-plugin).

If we want to add these, we'll have to do it in subsequent reviews.

Patch-set: 7
Label: Code-Review=+2