Run branch jobs on branchless CI code
In upstream we run stable branch jobs on master for CI related code
Because we don't have branches in CI code, we can break CI for
a stable branch unintentionally because we don't gate this.

Add jobs that will run on supported stable branches for CI related
repositories only - tripleo-ci, tripleo-quickstart(-extras).

Change-Id: I10872e2e7d50546773f8bba64abc0bb575b1f1e0