Update patch set 1
Patch Set 1:

> For consistency with other's config, I'd add the users that need to
> do the push to the ptl groups. Being owner of the repo they can do
> the direct push.

The context here is that we're trying to import the old github repo [1] here. We'd like to keep the history if possible, so when we push we get a "invalid committer email" error, because some of that history is commits by other folks. IIUC this patch should allow us to do that push. Honestly we can even revert this patch once the push is done.

[1] https://github.com/redhat-openstack/whitebox-tempest-plugin

Patch-set: 1