Leave only quickstart cloned in CI
In upstream we stopped to clone all projects in CI, do the same
in RDO cloud jobs. We don't need these projects to be cloned for
building them, buld-test-packages role clones them in case it
needs to build them.
It also solves problem with old branches that don't present in
some of projects. For example aodh doesn't have newton branch
already and it fails all gate newton jobs.

Change-Id: Ifa5032d3f68c6450c0f58168cbd6b841075ec7de