Automate GitHub repo creation for new project distgits
We've had partial support for automating new projects creation from
rdoinfo changes for a while. One of the missing pieces is the ability
to create the GitHub mirror for new distgits when new projects are
created, and this commit adds it via a post job for the config repo.

This job will check for newly created files in the resources/ directory,
then find any new distgit project, and create its mirror in GitHub
using an API key.

Change-Id: I330d052237382cb0918e48d4c879b725870e920f