Add new non-voting job to validate fedora rebuilds for deps
As part of the dependencies automation in RDO, a new job in rdoinfo check
pipeline will detects when a change happens in parameter
buildsys-tags: cloud7-openstack-<release>-candidate for a package with
config "fedora-dependency". In this case the job looks for a fedora
build with same nvr (except the dist part) as the specified in the tag
and performs a scratch rebuild in CBS using the srpm from fedora.

For example,  when following change is detected in rdoinfo:

--- a/deps.yml
+++ b/deps.yml
@@ -845,7 +845,7 @@ packages:
   buildsys-tags:
     cloud7-openstack-common-testing: python-django-babel-0.4.0-2.el7
     cloud7-openstack-common-release: python-django-babel-0.4.0-2.el7
     cloud7-openstack-queens-testing: python-django-babel-0.4.0-2.el7
+    cloud7-openstack-queens-candidate: python-django-babel-0.6.2-1.el7

this job will fetch srpm for python-django-babel-0.6.2-1.fc28 from
fedora and will do a scratch cbs build using target
cloud7-openstack-queens-el7.

In this phase the new job is non-voting. Once it's confirmed that it
works as expected it will be moved to voting and a new job to do the
actual build in CBS will be created.

Change-Id: I182569a11ab34e9a448f70d4fee806bafc28a62c