Lower the amount of min-ready nodes from 25 to 11
Jenkins is currently struggling to keep up with the concurrent
connections when running at -- or near -- full capacity. This is
likely one of the causes behind the slave disconnections we have been
seeing.

Between all the different nodes that we have, we can currently have
up to 25 idle nodes ("ready" in nodepool) to which jenkins must also
maintain a persistent connection which does not help.

Lowering the amount of min-ready nodes will impact the time it takes to
start a job but if we have less idle nodes, Jenkins should hopefully
be more reliable and it's an acceptable compromise.

Change-Id: Ib85418c77afaecb9db483bf68ab8e783512a1b51