Add dlrn_hash_tag and revert to default var names for images URLs
This patch partially reverts https://review.rdoproject.org/r/#/c/11528/

We need to keep variable names overcloud_image_url and ipa_image_url as
they are used in [1]. Additionally, those variables are used from
ovb-rdocloud.yml file in tripleo-ci repo [2] which overrides the value in
the release config. This file uses parameter dlrn_hash_tag in the value
for the url parameters so we need to define it in the custom
configuration.

[1] https://github.com/openstack/tripleo-quickstart-extras/blob/master/roles/overcloud-prep-images/templates/overcloud-prep-images.sh.j2
[2] https://github.com/openstack-infra/tripleo-ci/blob/master/toci-quickstart/config/testenv/ovb-rdocloud.yml

Change-Id: I1dd5a6aacdfdaf3c942362e19d72873f43381039