Update patch set 3
Patch Set 3: Code-Review+2

I created the web page, so it had to be there :). 

I think that group was added in the Gerrit interface, but not in the resources/config.yaml file, so that's why it was overwritten after the latest changes.

Patch-set: 3
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+2