Use non-containerized undercloud deployment for master in rdoinfo gate
Currently, we are using a non-containerized ovb job, but the config file
has been deleted so it's failing in the gate.

As per TripleO team feedback non-containerized overcloud is not longer
supported so we need to figure out how to gate package updates in
tripleo. Until then, let's start testing updates using undercloud
deployment which should work fine.

Change-Id: I281a6f46241e25a44575cdb4f536f20616152dd7