Increase precedence for check pipeline to normal
Currently, precedence is low and precedence for openstack-check is
normal, so jobs in check pipeline become queue for long time when
compete for nodes with jobs in openstack-check pipeline.

Note that i'm keeping precedence normal for gate pipeline, so this
change may increase queue time for jobs on it. If this is a problem
in future we could increase precedence to high for gate.

Change-Id: Icf27685b014c7e1505a92b9882e418d8184e490d