Update patch set 1
Patch Set 1: Code-Review-1

I thought the reason for this group naming is not that the jobs are branchless, but that we want to run this set of jobs on branchless projects like tqe, etc.

So I think the naming is correct, but we can debate it.

Patch-set: 1
Reviewer: Gerrit User 163 <163@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1