Add option to disable override robots in wget when getting from log
We do a recursive wget to fetch the created RPMs in a parent job from
the log server. Some time ago it was implemented a way to avoid robots
from indexing content from logs.rdoproject.org so now no RPMs are
fetched from the log server.

This patch adds an option "-e robots=off" to wget to override this
option.

Change-Id: I797b0cd720df01af08069ee97c3042931d0d7613