Update patch set 1
Patch Set 1:

I'm more in favor of removing the validation of the tags from the job. It adds time for the job to run, specially when doing massive tagging and since the creation of this jobs, we've never observed any issue wit it, so i really think we don't need to do this validation after tagging.

Patch-set: 1
Label: Workflow=0