Use test data for rdoinfra playbooks
Previously, we used the same Ansible vault file for production and
CI jobs. It's not a good idea, in case the vault password is compromised,
so for CI jobs we are now replacing the file with another one
containing test data and self-signed certificates.

This file will be included in-repo.

Change-Id: Ic8dd68d150cddf56f31151a241df1e920af94450