Update patch set 1
Patch Set 1: Code-Review+2

> I wonder why this package is special; other packages have empty
 > __init__.py too.

This is becuase of the way the rpmlint job works: it compares the number of errors rpmlint shows on the current package with the number of errors in the newly built package, and fails if the new one has more errors.

So, it the current package already has a zero-length file, there is no job failure.

Patch-set: 1
Label: Code-Review=+2