tripleo: remove useless skip-if
In a previous patches, we tried to configure negative regex in a skip-if
but looking at the code of zuul I don't think we can do that easily with
the version of zuul that is deployed in RDO CI.

Indeed, skip-if is used to give a list of files that if touched, the job
won't be executed.

In the case of fs042, we had all-files-match-any:
- ^extraconfig/services/(?!ipsec.yaml).*$
- config/general_config/featureset(?!042).*

but we rather need to say:
"when patching extraconfig/services/ipsec.* or
config/general_config/featureset042., kickoff the job". Which can be
done with "files:" layout.

For featureset011, 037 and 040, I haven't found any good layout that
will prevent to not run the jobs for nothing. If we use "files:", it'll
never run the jobs when we need it so we better keep it like this for
now.

For featureset001 and 035, same thing, the existing rules already
prevent to run the job on some irrelevant files, config/general_config/featureset(?!001|035).*
has no effect in that case, let's remove it.

Change-Id: I35caaa020e9f845935d095199d847926043c4fd5