Use local project list for upstream-centos-7 nodepool builds
Previously, we were pinning the list of OpenStack projects to
retrieve during the upstream-centos-7 build to a specific commit [1],
since we had to stop syncing our repo from upstream (Zuul v3 changes).
Now, some of those projects have been removed, so we need to keep
that list of projects locally.

[1] - https://git.openstack.org/cgit/openstack-infra/project-config/plain/gerrit/projects.yaml?id=d2efe2e71ed07e17b8f0922934665dfb3a6fdbd0

Change-Id: I5f7c26abbf415d273e56de4626ccc733e5ad49c0