Temporary remove oooq jobs for queens stable release
TripleO jobs for queens use a node upstream-centos-7 which takes
a long time to be available as it competes con ovb jobs for tripleo-ci.
This keeps all rdoinfo jobs in queue for more that 5 hours in the best
case which makes totally impractical.

This patch is removing queens jobs from gate. Note that these jobs
are only relevant for updates in CloudSIG for queens. We may temporarily
add them back when we have jobs updating it.

Change-Id: I96b08f0d9299860260377fe71eafc3cab8ba2e0b