add fs037/updates job to the periodic pipeline
Remove the -nv suffix for updates/fs037
Add it to the master and queens periodic job.
Initially update/fs037 should only be criteria for
master to promote.  Queens is starting to show
signs of life, so it doesn't hurt to run there as well
as long as it's not added to criteria.

Change-Id: Ib66a64649a63232725b92d71005b90973fa010b1