Provide a hook to add an additional SSH key in the rdo-base element
Jobs ran from jenkins on ci.centos.org will start using nodepool images
but the problem is that they are set up to have the public key of the
nodepool keypair.

The nodepool private ssh key isn't going to be set up on the jenkins
nodes so let's instead add the public key of the jenkins instead.

Change-Id: Iba95d1d6818d5101c6cf02d0b9fb4cb708972957