Switch the DLRN-rpmbuild-fedora job to use the fedora stable repos
The new Fedora 28 stable repo has been bootstrapped, and we are
planning to remove the Rawhide builder due to its instability. This
commit switches the DLRN-rpmbuild-fedora job to use the new stable
repos.

Please note that this commit depends on [1] to be merged to get
the right tag in use for the job.

[1] - https://softwarefactory-project.io/r/12186

Change-Id: I320ec5613f063a49176e0d763991a5d0195e287e