Add rocky-py3-uc tag update to rdoinfo upper-constraints update
Since we now have two upper-constraints-based tags for master
(rocky-uc for the main CentOS builder, and rocky-py3-uc for the
Fedora 28 stable builder), let's update the
rdoinfo-upper-constraints-update-master job to update both tags at
the same time.

Note that currently we do not have a CI job to test the updates
at the Fedora level, we will keep on gating with the centos-based
job.

Change-Id: I76b761d611627d927ba0bae5c305bd2264b6bfd6