Replace sf-zuul-worker element with zuul-cloner-pip in rdo-centos-7
The sf-zuul-worker element requires the centos-release-openstack-pike
package, and since it is installed it creates some package conflicts
for jobs using Newton or Ocata.

We don't need the zuul-swift-upload binary in the image because it is
not used, so we should be able to use zuul-cloner-pip, just like
we have in the fedora image.

Change-Id: Iadac9fd84a72d13d14c12865e87e3e557f86b262