Install pip packages in a virtualenv for rdoinfo jobs
With the latest image, we have pip 10.0.1. This version is known to
break the previous behavior of uninstalling setuptools-based packages
(like virtualenv or pyOpenSSL) [1]. Since some of these packages are
included in the image, trying to "sudo pip install" some packages
(e.g. rdopkg) fails, aborting jobs.

To prevent this, let's install everything in a virtualenv. For now
we are changing this for rdoinfo jobs, and we will extend to other
jobs if needed.

[1] - https://github.com/pypa/pip/issues/4805

Change-Id: I66cf30627e28c461ac45293ae09069bfb7fd734a