Add functional jobs to sinkhole pipelines
[1] is adding functional tests to sinkhole project. This
patch is creating jobs to run those tests both in centos
and fedora to ensure it works as expected in both cases.

Initially, we are adding them as non-voting, however we'll
move them to voting once we check they work fine.

Additionally, this patch is fixing sinkhole command to
provide mandatory subcommand according to change in [1]
to support python2.

[1] https://review.rdoproject.org/r/#/c/13974/

Change-Id: I2ca480b3f0fdffc01bdad5179dbb03af711d9dc7