Fix add-project-from-rdoinfo for Zuul v3
We need to change the python script to create the right lines for
Zuul v3, instead of the previous Zuul v2 configuration.

Since the process now requires creating three different reviews
(2 on review.rdo and one on softwarefactory-project.io), this commit
creates the two reviews in review.rdo, and informs that the third
review is required.

Note that this script will sort zuul.d/projects.yaml using the project
name as a key, so if projects are added but unsorted, an automated
review will move them around. Also, it tries to preserve the existing
indentation, but a first review will change it, since it is not 100%
consistent.

Change-Id: If2030b4df575e9cac4b1a303295cce8344d1279c
TG-1687