Fix add-project-from-rdoinfo for Zuul v3
We need to change the python script to create the right lines for
Zuul v3, instead of the previous Zuul v2 configuration.

Note that this script will sort zuul.d/projects.yaml using the project
name as a key, so if projects are added but unsorted, an automated
review will move them around.

Change-Id: If2030b4df575e9cac4b1a303295cce8344d1279c