Update patch set 6
Patch Set 6:

> > I see there are no secrets required by the jobs, so I think this
 > > should be in the rdo-jobs repo instead of the config repo.
 > 
 > Everything should be moved to rdo-jobs? or only job definitions? Is
 > there some guide about zuulv3 in RDO(all repositories? what to add
 > where?)?

Ok, now I have been able to spend enough time to understand the review (sorry I was in the middle of some meetings).

The goal is to create a 3rd party CI job for openstack/puppet-openstack-integration, IIUC. Then, we would have to:

- Create a new review to the rdo-jobs repo, that includes all the playbooks and zuul.d/jobs.yaml changes. This should go to that repo because we do not need to access any secrets, so it is better to use an untrusted project (rdo-jobs) than a trusted one (config).

- Keep this review, only with the zuul.d/projects.yaml addition, and make it Depends-On the rdo-jobs change.

Once all is merged, since this is in the experimental pipeline you will need to use the "check rdo experimental" keyword to trigger the jobs.

Patch-set: 6
Label: Code-Review=0