[WIP] create a base job to replace te-broker infra server
the new job will use nodepool tenant credential in rdo cloud to create a
test environment to run OVB jobs, using existing methods currently used
in the job reproduction process (ovb-manage-stack role)
This is just a proof of concept patch, with no intention to be final.
Feedback that I need from this patch:
1) is it acceptable to have this playbook here ?
2) is it acceptablt to have this new role here ?
3) the playbook and role will download and use an external repository
(openstack-virtual-baremetal) to effectively create the OVB test
environment. Is this acceptable ?

Change-Id: I1e5825e1a8fccd1b1674fcede2d6e08fcdefa449