Update patch set 1
Patch Set 1: Code-Review-1

This is not deleting the keys from the ssh-agent session created in a previous task but deleting the keys from the session just created which is empty. You need to define variables SSH_AUTH_SOCK and SSH_AGENT_PID instead of creating a new ssh-agent.

Patch-set: 1
Label: Code-Review=-1