Parent base-test to base-minimal-test
This is to validate our new base-minimal jobs actually work. Also, we
don't actually need base-test to live in config project anymore, once
we confirm the parent works as expected, the follow up patch will
delete base-test.  Since base will eventually move to rdo-jobs and
we'll be able to use speculative for it.

Change-Id: I932c987f12b563d6bf7abcb3cbf13330a39845e0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>