Add ykarel as provenpackager
In the last months, Yatin Karel (ykarel on irc) has been doing
a great job in maintaining packages in RDO, sending fixes and
reviewing other's reviews.

I think it's time to add him to the provenpackagers group so that
he becomes core contributor to all distgits in RDO.

Thanks ykarel for such a great job!

Change-Id: If611d02e2521bbac81b31010cf7beb86cc99ae9b