Add rocky to rdo-send-stable-release job
This will automatically create reviews for new releases created in
upstream releases projects in rocky. Note that the script has a check
to find out if rocky-rdo branch exists, so no reviews will be created
for unbranched projects.

It also removes newton as there are no more releases for newton.

Change-Id: Ic2e46a04fa1bef29a3d53927ff5ec4ad84c0e5f7