Update patch set 2
Patch Set 2:

Please note that this is also going to affect some non-TripleO projects (at least rpm-packaging and puppet-openstack-integration).

I'm not against it, as long as we keep the shutdown time to a minimum (1-2 days, as stated in the e-mail).

Patch-set: 2
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0