Create distgit for openvswitch
In order to automate and have peer-review for new builds of
openvswitch in RDO deps repo we need to have a distgit repo
managed by the gerrit instance in RDO.

This patch is creating the distgit. Once created I'll create
branches with the existing content in [1] although I'll rename
the branches to align with OpenStack releases instead of ovs
versions (we'll need it to automate builds).

Note this is just a first attempt and will need to experiment
a bit to get the better workflow and align RDO rebuilds with
fedora ones.

[1] https://github.com/rdo-common/openvswitch

Change-Id: I2d8aa8e50788125710f92638cbea814e9fe0b353