Add f28 promotion job containers-build
With the premise of failing at the beginning, the job
**periodic-tripleo-fedora-28-master-containers-build**
depends on promote-consistent-to-tripleo-ci-testing job to run.

kolla-build settings that are being changed:
 * kolla_base
 * kolla_base_image
 * kolla_base_tag
 * kolla_virtualenv
 * kolla_push
 * trunk_repository
 * trunk_deps_repository

Manually tested this change to make sure it does not break
centos7 containers build process. See:
https://tree.taiga.io/project/tripleo-ci-board/task/567

The job won't build any container until f28 support merges
in kolla. When this happen it will be not pushing anything
to rdo trunk registry, so it is a safe move.

Story: https://tree.taiga.io/project/tripleo-ci-board/us/447

Change-Id: I53682b45db9b1c4b38afc55e8ba677892c4dbfff