WIP: Add pure f28 promotion jobs (bm/containers) to periodic pipeline
Note: it was agreed that the pure fedora-28 jobs would exist in the
openstack-periodic pipeline alongside the centos-7 jobs.

 - featureset002-upload
 - Note: f28 containers-build is defined in project: config [1](trusted)
 - standalone-f28

No dependencies defined as the intent is to monitor the progress
on all the steps on f28 pipeline, as the containers build job.

These jobs can be added to the periodic pipeline only after merging
the following job definitions in the untrusted rdo-jobs repo.

[1] https://review.rdoproject.org/r/#/c/17738

Depends-On: https://review.rdoproject.org/r/17739
Depends-On: https://review.rdoproject.org/r/17807
# Add depends-on that points to rlandy's f28 release file.

Change-Id: I9f4132f6d7d4a389465a2b7612b7fade7edef2ae