Update patch set 10
Patch Set 10: Code-Review+2

(1 comment)

OK - so for the purpose of JUST building containers, this looks right.
In so far as https://review.rdoproject.org/r/#/c/17744/11/playbooks/tripleo-ci-base-promote-consistent-to-tripleo-ci-testing/run.yaml is ruses the same job to promotes both centos7 and fedora.
So we are not at a pure fedora promote to container build yet.

Patch-set: 10