Update patch set 1
Patch Set 1:

fedora mixed job is dependent on the consistent promotion job. If that fails the job is skipped, if it succeed, the job is run. If the mixed job runs and the hash didn't promote, it would test an old tripleo-ci-testing hash. If the mixed job is a criteria for the centos promotion, the run will not promote anyway. If it's not then we are promoting a centos hash that may not work with a fedora hash.
In the design discussion we opted for the fixed hash approach, and decided centos and fedora would promote together, or not promote.

Patch-set: 1
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0