promoting consistent: don't fail the job if fedora hash is not found
We don't want to skip the entire execution if fedora and centos are not
in sync.
If fedora equivalent is not found, the fedora job will run anyway, but
will test *AND REPORT* on the old tripleo-ci-testing hash, so since the
fedora job is part of the criteria for promotion by default, the hash
will not promote without manual intervention, but we'll get feedback on
the centos jobs.

Change-Id: I5c2df4b146233f102626fdf6d3b62889d434d9f0