Delete CREATE_FAILED stacks
When we have CREATE_FAILED stack we fail on task of getting servers
and job never gets to the step of deleting stack.
So let's ignore all work *before* deleting stack, because all or
part of these resources could be missing. And delete the stack
anyway - not depending on tasks status bedore.

Change-Id: I2480d9c89ff19184135dc5c1bac43159d7bd2c37