Assing jobs to the right files in rdoinfo after reorganization
After rdoinfo reorganization we have per-release and per-tag files.
Let's take advantadge of this and assign the jobs for each release to
the files which actually modify it.

This patch also moves some experimental jobs to the check pipeline as
they will run only when they need to. This improves dependencies and
packages updates gating.

Change-Id: I32364bc3a56a9d78b7e27764029446be7885d7d2