Add new periodic containers-build-push inheriting from upstream
As tracked in [1] we want the periodic containers build job to use
the same approach as the upstream build-containers-centos-7. This
adds periodic-tripleo-centos-7-master-containers-build-push
with an also new job as its parent
periodic-tripleo-ci-base-singlenode-rdo-containers-build-push
which then inherits from upstream tipleo-build-containers-centos-7.

The new job also handles login to the registry so we can use
the existing rdo registry token. See [2] for the related change in
the upstream job to accept parameters passed from here.

[1] https://tree.taiga.io/project/tripleo-ci-board/us/652
Change-Id: I93519f42a640ece6dcbb2e2f1d85f8180577102c