Update patch set 5
Patch Set 5: Code-Review+1

This should be safe to merge. Noop, not activated anywhere but ready to be tested by adding a post-run: pre-run: value that will just duplicate the operation. I suggest to start from the reporting, which is safer. If the setup messes with the files, it could be harmful.

Patch-set: 5
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1