Add new f28 job and repo override+push in centos containers build
Adds the buildcontainers_override_repos set to tripleo-ci-testing
for the periodic version of the centos build containers job.
This depends on /#/c/638652/ which adds support for the override
in the upstream version from which this periodic inherits. Tracked
in [1] by tripleo-ci squad -  for more context behind this change

Enable push in periodic-tripleo-centos-7 containers-build-push job
The periodic-tripleo-centos-7-master-containers-build-push added
in https://review.rdoproject.org/r/18729 has push_containers false
Now that we've seen the job running lets also try to push to rdo
registry with "ci-squad-testing". Tracked by ci squad in [2]

Finally add periodic f28 containers-build-push based on upstream

As tracked in [3] we want the periodic f28 containers build job
to use the same approach as the upstream build-containers-centos7.
Adds new base periodic-tripleo-master-containers-build-push-base
and we use it for both centos-7-master-containers-build-push and
the new fedora-28 job added here
periodic-tripleo-fedora-28-master-containers-build-push

Support for py3 packages and f28 support for the containers build
is added in [4] and [5] adds support for tripleo-ci-testing repos

[1] https://tree.taiga.io/project/tripleo-ci-board/task/773
[2] https://tree.taiga.io/project/tripleo-ci-board/task/817
[3] https://tree.taiga.io/project/tripleo-ci-board/us/651
[4] https://review.openstack.org/636160
[5] https://review.openstack.org/638652
Change-Id: Ie01fe59ec1725a252c3d58e766dfaf4cd791f210