Add periodic f28 containers-build-push based on the upstream
As tracked in [1] we want the periodic f28 containers build job
to use the same approach as the upstream build-containers-centos7.

Adds new base periodic-tripleo-master-containers-build-push-base
and we use it for existing centos-7-master-containers-build-push
and the new fedora-28 job added here
periodic-tripleo-fedora-28-master-containers-build-push

Support for py3 packages and f28 support for the containers build
is added in the Depends-On.

Depends-On: https://review.openstack.org/638652
[1] https://tree.taiga.io/project/tripleo-ci-board/us/651

Change-Id: If30539872ace760e82374b97888a6f81dd055375