Update patch set 2
Patch Set 2: Code-Review-1

I am worried that this would force us to make blindfolded maintenance in the future. I see whitelists and other config items there for which I want to be able to see build results before merging a change.

That is my personal view, but I will respect whaever the rest of the team ops in for, i am just trying to avoid making out life harder.

Patch-set: 2