Make containers-build-push-base job branchless
Simplify base job so we can override release in the child job
and avoid duplicating config for all branches.

Create a new job definition without -master that will be the new
base job for all releases. Temporarily dup job definitions
that will be reparented and removed in a follow-up patch.

Change-Id: Ia135e96821d058ea6ca7779706cdeca399dbd049