Creates new branch rocky-rdo
The ansible-role-openstack-operations is needed before the upgrades, as
it is going to be use to do the backup of the undercloud and the
control-plane.
The idea is to deliver it on the Stein release so, before doing the
upgrade a backup of the nodes can be done.
To do the backup of the Rocky nodes the package needs to be
deliver on the Rocky release to avoid possible undesired packages
updates.

Change-Id: I9b92c661e9fc90666b4052c6e8e228a9e3d49bbe