Switch to buildah on container build jobs
Switch from docker to buildah on master/stein container build jobs
in the periodic pipeline. This change modifies the layout for the
jobs defined in the promotion pipeline, that is, old jobs named
as -build-push for master and stein branches are replaced with
-buildah suffix.

The jobs for rocky and older branches, docker remains with the
name -build-push suffix.

Jobs are defined in the trusted config at [1] and have been tested
in testproject [2].

[1] https://review.rdoproject.org/r/#/c/20000/
[2] https://review.rdoproject.org/r/#/c/20023/

Story: https://tree.taiga.io/project/tripleo-ci-board/us/815

Task #964: Add container build buildah jobs to promotion pipeline
(layout).

Change-Id: I48b4669dc8498bfbd844a6b705dcef5d12774657