Update patch set 1
Patch Set 1:

Dmitry, this is third party non-gating.  There should not be a high barrier to entry for 3rd party jobs.  Yes we can socialize this a bit more however the project itself can not prevent this from running via third party, that is what 3rd party is for.

Requirements for 3rd party:
https://docs.openstack.org/infra/system-config/third_party.html

Patch-set: 1