Update patch set 1
Patch Set 1:

The main problem I'm talking about finding a cause of a failure without knowing the exactly way CI puts files. Especially since this layout is very different from devstack's (which is not anyone's fault, but people are used to devstack, soo..). Take for example the fact that the most useful things are under logs/undercloud/home/zuul. The undercloud bit is already not entirely obvious to a newcomer, but home is pretty insane :)

Patch-set: 1