Update patch set 1
Patch Set 1:

>  the current logs layout is same for all tripleo jobs

Sure. But you're adding a job to a project that is not a TripleO project, and people working on it are mostly not involved in TripleO.

Let me probably ask this: what exactly result are you planning to achieve with this job? If this is job for tracking which change broke what, I withdraw my objections. If you hope that the ironic team will use the job to avoid breaking TripleO, it may be overly optimistic.

Patch-set: 1