run docker login by default in every rdo jobs.
We need all the jobs to login into docker or podman with credentials.
This change installs docker and podman, logs in and remove both
packages.
The resulting auth file with token credentials can be reused at any
time, and the change doesn't fail if any problem is encountered, the
rescue task will just print a warning message

Change-Id: I692d6dc4a08ecc60c9744c0f985b6a790e8c205b