[WIP] global login for docker and podman, with attempt at testing before merging
Adds tasks to login to registry via podman and docker by default, and
tests.
The set of tasks are destined to run in ALL the jobs, and can
potentially break every job we have, so to test it before merging it,
I'm trying a new method:
- clone all the files my change would touch
- clone all the job my change is involved in
- add the changes to those files
- merge everything. As changes are merged in definition that are not
used in production, merging is safe.
- launch staging check jobs b triggering changes in testproject
- when we are sure that the cloned jobs are working, merge the change to
the original file, then merge the change.

At this moment, I have no intention of runnig a whole end to end
integration test just to be sure that a set of 2 tasks works correctly,
so I created a functional test taht is launched in the cloned job,
feedback will come much faster this way, and we leave trace for all the
things we now know htye may represent a problem in the future

Change-Id: I3edfaf4427aa008ef3f0494f144884ef6c5f2a7f