Adds empty molecule job
Allows us to run molecule job as non-voting when making changes to
roles inside the repository.

In order to avoid pytest execution failure when no tests are discovered,
this change contains one empty molecule test definition that does
nothing. The execution takes ~2s to run because it does no provisioning.

Change-Id: Ibd38a1dc94faeb27369da528d7b69b0eaa0dea68