Update patch set 9
Patch Set 9: Code-Review-2

(3 comments)

I don't continue to review here, this patch is turned to be completely out of scope of the task. It should install openstack deps according to distro, it has nothing common with molecule testing. Anything related to molecule should be in another task, another patch and not to be mentioned or noticed here.

Patch-set: 9
Label: Code-Review=-2